# Detect function usages