--- a/src/SymbolKind.php
+++ b/src/SymbolKind.php
@@ -1,0 +1,12 @@
+<?php declare(strict_types = 1);
+
+namespace ShipMonk\ComposerDependencyAnalyser;
+
+final class SymbolKind
+{
+
+    public const CLASSLIKE = 1;
+    public const FUNCTION = 2;
+    public const CONSTANT = 3;
+
+}