… README.md
… src/Config/Ignore/UnusedClassIgnore.php
… src/Result/ConsoleFormatter.php
… src/Result/JunitFormatter.php
… src/UsedSymbolExtractor.php
… tests/AnalyserTest.php
… tests/ConfigurationTest.php
… tests/ConsoleFormatterTest.php
… tests/InitializerTest.php
… tests/JunitFormatterTest.php
… tests/UsedSymbolExtractorTest.php
… tests/data/not-autoloaded/functions/fn-usage.php
… tests/data/not-autoloaded/functions/org/package/fn-def.php
… tests/data/not-autoloaded/used-symbols/other-symbols.php
… tests/data/not-autoloaded/used-symbols/use-statements.php