Resolves #67