@janedbal 2024-04-08
No measurable performance impact:

```
# before
Benchmark 1: vendor/bin/composer-dependency-analyser
  Time (mean ± σ):      1.448 s ±  0.041 s    [User: 1.243 s, System: 0.203 s]
  Range (min … max):    1.406 s …  1.553 s    10 runs

# after
Benchmark 1: vendor/bin/composer-dependency-analyser
  Time (mean ± σ):      1.448 s ±  0.019 s    [User: 1.253 s, System: 0.193 s]
  Range (min … max):    1.419 s …  1.474 s    10 runs


```