@jdreesen 2024-08-30
Should this detect usages of `trigger_deprecation()` from `symfony/deprecation-contracts`?
Because for me, it didn't work.