@JanTvrdik 2024-08-30
It does not work, because it's a global function referenced with neither FQN nor use statement, i.e. both the following would work.

```php
use function trigger_deprecation;

trigger_deprecation(
    'teamneusta/pimcore-testing-framework',
    '0.13',
    'The "%s" class is deprecated, use "%s" instead.',
    TestKernel::class,
    RootTestKernel::class,
);
```

or 


```php
\trigger_deprecation(
    'teamneusta/pimcore-testing-framework',
    '0.13',
    'The "%s" class is deprecated, use "%s" instead.',
    TestKernel::class,
    RootTestKernel::class,
);
```