@jdreesen 2024-08-30
I see. I'll try that when I have some time again. Thx for the quick reply :)