@jdreesen 2024-09-04
I just tried prepending the `\`, but it still says `symfony/deprecation-contracts` is unused. 
But now it also detects `symfony/contracts` as a shadow dependency.

Without leading `\`:
![grafik](https://github.com/user-attachments/assets/46c32c69-1e95-48f3-81ab-a30af4692657)

With leading `\`:
![grafik](https://github.com/user-attachments/assets/0e1ac08f-5935-41b4-ad79-44b35bc8b000)