@janedbal 2024-09-05
It just means that `symfony/contracts` is the actual library where the function [gets declared first](https://github.com/symfony/contracts/blob/55d40d0f56bbcbf10fb51048599815cc6f6b2bd4/Deprecation/function.php#L23).