@jdreesen 2024-09-05
It's probably because one of my dependencies has `symfony/contracts` as its dependency, and `symfony/contracts` says [it replaces `symfony/deprecation-contracts`](https://github.com/symfony/contracts/blob/f8cd5313753cfac8329ebc4033e2013b874208e1/composer.json#L32).